test(wgpu): share the texture readback between the presentable-root and clear contracts
